changed project layout moving main out of src folder
