+++ .gitignore
@@ -13,3 +13,5 @@
 
 config.json
 output/
+
+__pycache__/
