+++ src/utils.py
@@ -57,4 +57,7 @@ def delete_temp_files(directory=check_config("TEMPFILEPATH")):
         file_path = os.path.join(directory, file)
         if os.path.isfile(file_path):
             os.remove(file_path)
-            print(f"Deleted temporary file: {file_path.encode("utf-8")}")
+            print(f"Deleted temporary file: {file_path.encode('utf-8')}")
+
+def check_directory():
+    run_command('cd')
